+++ esm/index.js
@@ -1,0 +1,8 @@
+/* eslint-disable node/no-unpublished-import, node/no-unsupported-features/es-syntax, import/extensions */
+
+import MarkdownItRenderer from '../src/index.js';
+
+export default MarkdownItRenderer;
+
+export { default as markdownItPlugin } from '../src/markdown-it-plugin';
+export { default as MarkdownItRenderer } from '../src/index.js';