+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 All notable changes to this project will be documented in this file. This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [4.0.0](https://github.com/uttori/uttori-plugin-renderer-markdown-it/compare/v3.2.3...v4.0.0) - 2020-01-16
+
+- 🧰 Add ESM Support
+- 🧰 Add explicit exports
+- 🧰 Support for [WikiLinks](https://en.wikipedia.org/wiki/Help:Link#Wikilinks_(internal_links)) style linking
+- 🧰 Support for generating a Table of Contents based on `H#` header tags with `[toc]`
+- 🛠 Fix config extension with custom options
+- 🎁 Update dev dependencies
+
 ## [3.2.5](https://github.com/uttori/uttori-plugin-renderer-markdown-it/compare/v3.2.3...v3.2.4) - 2020-01-02
 
 - 🎁 Update dev dependencies